requirements.txt
# Core GUI Framework
customtkinter>=5.2.0

# AI and Machine Learning
torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
faster-whisper>=0.10.0
openai>=1.3.0
python-dotenv>=1.0.0

# Audio Processing
sounddevice>=0.4.6
simpleaudio>=1.0.4
webrtcvad>=2.0.10
numpy>=1.24.0

# TTS (Text-to-Speech)
kokoro-tts>=0.1.0

# Language Detection
langdetect>=1.0.9

# Async and Threading
asyncio-mqtt>=0.13.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: For better audio quality
pyaudio>=0.2.11
